Move future event handling to separate file
