+++ backend/redis/events.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"strconv"
 
 	"github.com/cschleiden/go-workflows/backend/history"
 	"github.com/cschleiden/go-workflows/core"
@@ -103,59 +102,4 @@ func addEventsToStreamP(ctx context.Context, p redis.Pipeliner, streamKey string
 	addEventsToStreamCmd.Run(ctx, p, []string{streamKey}, eventsData)
 
 	return nil
-}
-
-// Adds an event to be delivered in the future. Not cluster-safe.
-// KEYS[1] - future event zset key
-// KEYS[2] - future event key
-// KEYS[3] - instance payload key
-// ARGV[1] - timestamp
-// ARGV[2] - Instance segment
-// ARGV[3] - event id
-// ARGV[4] - event data
-// ARGV[5] - event payload
-var addFutureEventCmd = redis.NewScript(`
-	redis.call("ZADD", KEYS[1], ARGV[1], KEYS[2])
-	redis.call("HSET", KEYS[2], "instance", ARGV[2], "id", ARGV[3], "event", ARGV[4])
-	redis.call("HSETNX", KEYS[3], ARGV[3], ARGV[5])
-	return 0
-`)
-
-func addFutureEventP(ctx context.Context, p redis.Pipeliner, instance *core.WorkflowInstance, event *history.Event) error {
-	eventData, err := marshalEventWithoutAttributes(event)
-	if err != nil {
-		return err
-	}
-
-	payloadEventData, err := json.Marshal(event.Attributes)
-	if err != nil {
-		return err
-	}
-
-	return addFutureEventCmd.Run(
-		ctx, p,
-		[]string{futureEventsKey(), futureEventKey(instance, event.ScheduleEventID), payloadKey(instance)},
-		strconv.FormatInt(event.VisibleAt.UnixMilli(), 10),
-		instanceSegment(instance),
-		event.ID,
-		string(eventData),
-		string(payloadEventData),
-	).Err()
-}
-
-// Remove a scheduled future event. Not cluster-safe.
-// KEYS[1] - future event zset key
-// KEYS[2] - future event key
-// KEYS[3] - instance payload key
-var removeFutureEventCmd = redis.NewScript(`
-	redis.call("ZREM", KEYS[1], KEYS[2])
-	local eventID = redis.call("HGET", KEYS[2], "id")
-	redis.call("HDEL", KEYS[3], eventID)
-	return redis.call("DEL", KEYS[2])
-`)
-
-// removeFutureEvent removes a scheduled future event for the given event. Events are associated via their ScheduleEventID
-func removeFutureEventP(ctx context.Context, p redis.Pipeliner, instance *core.WorkflowInstance, event *history.Event) {
-	key := futureEventKey(instance, event.ScheduleEventID)
-	removeFutureEventCmd.Run(ctx, p, []string{futureEventsKey(), key, payloadKey(instance)})
 }
