+++ backend/redis/keys.go
@@ -54,7 +54,7 @@ func futureEventsKey() string {
 }
 
 func futureEventKey(instance *core.WorkflowInstance, scheduleEventID int64) string {
-	return fmt.Sprintf("future-event:%v:%v:%v", instance.InstanceID, instance.ExecutionID, scheduleEventID)
+	return fmt.Sprintf("future-event:%v:%v", instanceSegment(instance), scheduleEventID)
 }
 
 func payloadKey(instance *core.WorkflowInstance) string {
