+++ backend/redis/workflow.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"strconv"
 	"time"
 
 	"github.com/cschleiden/go-workflows/backend"
@@ -17,56 +16,9 @@ import (
 	"go.opentelemetry.io/otel/trace"
 )
 
-// Find all due future events. For each event:
-// - Look up event data
-// - Add to pending event stream for workflow instance
-// - Try to queue workflow task for workflow instance
-// - Remove event from future event set and delete event data
-//
-// KEYS[1] - future event set key
-// KEYS[2] - workflow task queue stream
-// KEYS[3] - workflow task queue set
-// ARGV[1] - current timestamp for zrange
-//
-// Note: this does not work with Redis Cluster since not all keys are passed into the script.
-var futureEventsCmd = redis.NewScript(`
-	-- Find events which should become visible now
-	local events = redis.call("ZRANGE", KEYS[1], "-inf", ARGV[1], "BYSCORE")
-	for i = 1, #events do
-		local instanceSegment = redis.call("HGET", events[i], "instance")
-
-		-- Add event to pending event stream
-		local eventData = redis.call("HGET", events[i], "event")
-		local pending_events_key = "pending-events:" .. instanceSegment
-		redis.call("XADD", pending_events_key, "*", "event", eventData)
-
-		-- Try to queue workflow task
-		local already_queued = redis.call("SADD", KEYS[3], instanceSegment)
-		if already_queued ~= 0 then
-			redis.call("XADD", KEYS[2], "*", "id", instanceSegment, "data", "")
-		end
-
-		-- Delete event hash data
-		redis.call("DEL", events[i])
-		redis.call("ZREM", KEYS[1], events[i])
-	end
-
-	return #events
-`)
-
 func (rb *redisBackend) GetWorkflowTask(ctx context.Context) (*backend.WorkflowTask, error) {
-	// Check for future events
-	now := time.Now().UnixMilli()
-	nowStr := strconv.FormatInt(now, 10)
-
-	queueKeys := rb.workflowQueue.Keys()
-
-	if _, err := futureEventsCmd.Run(ctx, rb.rdb, []string{
-		futureEventsKey(),
-		queueKeys.StreamKey,
-		queueKeys.SetKey,
-	}, nowStr).Result(); err != nil && err != redis.Nil {
-		return nil, fmt.Errorf("checking future events: %w", err)
+	if err := scheduleFutureEvents(ctx, rb); err != nil {
+		return nil, fmt.Errorf("scheduling future events: %w", err)
 	}
 
 	// Try to get a workflow task, this locks the instance when it dequeues one
